Folder enumeration and destroy_vms
- Added function to enumerate folders and their objects (UNTESTED)

 destroy_vms:
- Added more checks
- Added path traversal
- Added folder structure output (UNTESTED)

+++ destroy_vms.py
@@ -30,12 +30,13 @@ Options:
 
 from docopt import docopt
 import logging
+from pprint import pprint
 
 from automation.utils import prompt_y_n_question, setup_logging, make_vsphere, warning, user_input, default_prompt
 from automation.vsphere.vm_utils import destroy_vm
-from automation.vsphere.vsphere_utils import cleanup
+from automation.vsphere.vsphere_utils import cleanup, enumerate_folder, traverse_path
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 
 args = docopt(__doc__, version=__version__, help=True)
 setup_logging(filename='destroy_vms.log', console_level=logging.DEBUG if args["--verbose"] else logging.INFO)
@@ -43,20 +44,32 @@ setup_logging(filename='destroy_vms.log', console_level=logging.DEBUG if args["-
 server = make_vsphere(args["--file"])
 warning()
 
-# TODO: prompt to destroy templates?
 if prompt_y_n_question("Do you wish to destroy a single VM? "):
-    vm, vm_name = user_input("Name of VM to destroy: ", "VM", server.get_vm)
-    logging.info("Destroying VM with name %s", vm_name)
-    destroy_vm(vm)
+    # TODO: VM at path + folder at path utility func
+    vm, vm_name = user_input("Name of or path to VM to destroy: ", "VM",
+                             lambda x: traverse_path(server.get_folder(), x) if '/' in x else server.get_vm(x))
+    if vm.config.template:  # Warn if template
+        if not prompt_y_n_question("VM %s is a Template. Do you wish to continue? " % vm_name):
+            exit(0)
+    if prompt_y_n_question("Continue with destruction? "):
+        logging.info("Destroying VM with name %s", vm_name)
+        destroy_vm(vm)
+    else:
+        logging.info("Destruction cancelled")
 else:
-    folder, folder_name = user_input("Name of folder to destroy: ", "folder", server.get_folder)
-    prefix = default_prompt("Do you wish to destroy VMs with a specific prefix? ", default=None)
-    recursive = prompt_y_n_question("Do you wish to recursively descend the folder tree? ")
-    destroy_folders = prompt_y_n_question("Do you wish to destroy folders in addition to VMs? ")
-    destroy_self = prompt_y_n_question("Do you wish to destroy the folder itself? ")
+    folder, fname = user_input("Name of or path to folder to destroy: ", "folder",
+                               lambda x: traverse_path(server.get_folder(), x) if '/' in x else server.get_folder(x))
+    pprint(enumerate_folder(folder))  # Display folder structure
 
-    logging.info("Carrying out destruction of folder %s with following arguments: "
-                 "Prefix: %s\tRecursive: %s\tFolder-destruction: %s\tSelf-destruction: %s",
-                 folder_name, str(prefix), str(recursive), str(destroy_folders), str(destroy_self))
+    prefix = default_prompt("What is the prefix of VMs you wish to destroy? Press Enter for none... ", default=None)
+    recursive = prompt_y_n_question("Recursively descend into folders? ")
+    destroy_folders = prompt_y_n_question("Destroy folders in addition to VMs? ")
+    destroy_self = prompt_y_n_question("Destroy the folder itself? ")
 
-    cleanup(folder, prefix=prefix, recursive=recursive, destroy_folders=destroy_folders, destroy_self=destroy_self)
+    logging.info("Destruction options - Prefix: %s\tRecursive: %s\tFolder-destruction: %s\tSelf-destruction: %s",
+                 fname, str(prefix), str(recursive), str(destroy_folders), str(destroy_self))
+    if prompt_y_n_question("Continue with destruction? "):
+        logging.info("Destroying...")
+        cleanup(folder, prefix=prefix, recursive=recursive, destroy_folders=destroy_folders, destroy_self=destroy_self)
+    else:
+        logging.info("Destruction cancelled")
